TI-2015 Move season definition to types packages and update types
