+++ deploy-scripts/package.json
@@ -17,7 +17,6 @@
   "author": "talent::digital",
   "license": "See license.md",
   "devDependencies": {
-    "@talentdigital/sdk": "workspace:*",
     "@types/node": "^18.7.18",
     "@vercel/ncc": "^0.34.0",
     "dotenv": "^16.0.1",
