+++ deploy-scripts/src/helpers/deploy-test-items.ts
@@ -1,5 +1,5 @@
 import got from "got";
-import { SeasonDefinition } from "@talentdigital/sdk";
+import { SeasonDefinition } from "types";
 
 type TestItem = {
   documentation: string;
@@ -58,7 +58,7 @@ const extractTestItems = (
 
             return {
               documentation: JSON.stringify(testItem.documentation),
-              eventType: testItem.id,
+              eventType: testItem.eventTypeId,
               level: levels[testItem.level],
               subCompetenceId: Number(subCompetenceKey),
               testId: testItemKey,
