+++ packages/sdk/src/index.ts
@@ -1,5 +1,3 @@
 import { TdSdk } from "./td-sdk";
 
-export * from "./season-definition/SeasonDefinition";
-
 export default TdSdk;
