+++ packages/types/SeasonDefinition.json
