+++ packages/types/SeasonDefinition.ts
