+++ packages/types/SeasonDefinitionTypes.json
