+++ packages/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./season";
+export * from "./SeasonDefinition";
